fix: 댓글 작성과 삭제 시 토큰 인증

+++ src/main/java/com/daedongmap/daedongmap/comment/service/CommentService.java
@@ -3,7 +3,6 @@ package com.daedongmap.daedongmap.comment.service;
 import com.daedongmap.daedongmap.comment.domain.Comment;
 import com.daedongmap.daedongmap.comment.dto.CommentDto;
 import com.daedongmap.daedongmap.comment.dto.CommentCreateDto;
-import com.daedongmap.daedongmap.comment.dto.CommentUpdateDto;
 import com.daedongmap.daedongmap.comment.dto.CommentWithRepliesDto;
 import com.daedongmap.daedongmap.comment.repository.CommentRepository;
 import com.daedongmap.daedongmap.exception.CustomException;
@@ -11,11 +10,9 @@ import com.daedongmap.daedongmap.exception.ErrorCode;
 import com.daedongmap.daedongmap.review.domain.Review;
 import com.daedongmap.daedongmap.review.repository.ReviewRepository;
 import com.daedongmap.daedongmap.user.domain.Users;
-import com.daedongmap.daedongmap.user.dto.UserBasicInfoDto;
 import com.daedongmap.daedongmap.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -32,8 +29,8 @@ public class CommentService {
     private final ReviewRepository reviewRepository;
 
     @Transactional
-    public CommentDto createComment(CommentCreateDto commentCreateDto) {
-        Users user = userRepository.findById(commentCreateDto.getUserId()).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+    public CommentDto createComment(Long userId, CommentCreateDto commentCreateDto) {
+        Users user = getUserById(userId);
         Review review = reviewRepository.findById(commentCreateDto.getReviewId()).orElseThrow(() -> new CustomException(ErrorCode.REVIEW_NOT_FOUND));
 
         Comment comment = Comment.builder()
@@ -72,19 +69,25 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentDto updateComment(Long commentId, CommentUpdateDto commentUpdateDto) {
-        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new CustomException(ErrorCode.COMMENT_NOT_FOUND));
-        comment.updateComment(commentUpdateDto);
-        return new CommentDto(comment);
+    public void deleteComment(Long commentId, Long userId) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new CustomException(ErrorCode.COMMENT_NOT_FOUND));
+
+        if (isCommentOwner(userId, comment)) {
+            throw new CustomException(ErrorCode.COMMENT_NOT_MINE);
+        }
+
+        List<Comment> replies = commentRepository.findAllByParentId(commentId);
+        commentRepository.deleteAll(replies);
+        commentRepository.deleteById(commentId);
     }
 
-    @Transactional
-    public void deleteComment(Long commentId) {
-        List<Comment> replies = commentRepository.findAllByParentId(commentId);
-        for (Comment reply : replies) {
-            commentRepository.deleteById(reply.getId());
-        }
-        commentRepository.deleteById(commentId);
+    private Users getUserById(Long userId) {
+        return userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+    }
+
+    private boolean isCommentOwner(Long userId, Comment comment) {
+        return comment.getUser().getId().equals(userId);
     }
 
 }
